Megre UIConstants to constants
