--- src/main/java/frc/robot/constants/UIConstants.java
@@ -1,17 +1,0 @@
-package frc.robot.constants;
-
-public final class UIConstants {
-    public static final class StartPositionUI {
-        public static final String POSITION_KEY = "StartPosition";
-        public static final String[] POSITIONS = {"Left", "Center", "Right"};
-        public static final double SELECTOR_WIDTH = 150;
-        public static final double SELECTOR_HEIGHT = 40;
-    }
-
-    public static final class CoralPositionUI {
-        public static final String POSITION_KEY = "CoralPosition";
-        public static final String[] POSITIONS = {"None", "Amp", "Source", "Stage"};
-        public static final double SELECTOR_WIDTH = 150;
-        public static final double SELECTOR_HEIGHT = 40;
-    }
-}
